:recycle: refactor: page 내에 기재한 cardSectionList를 CardList 컴포넌트로 분리
